Library Automation System - ASP.NET + MsSQL
This project is a library automation system using ASP.NET Core MVC and MsSQL. It includes admin and student roles. Admins can add, update, and delete books and categories. Students can borrow and view books but not modify them. Users can register and log in to perform tasks based on their roles. All actions are recorded in the database.